refactor(microvm)!: delete the agent-tunnel MicroVM path — the per-box gRPC tunnel, its pool/manager/workspace layer and the backend built on it, keeping only the worker-id encoding and client helpers lite still needs
